refactor(replication): extract resync contracts crate
