+++ Cargo.toml
@@ -39,6 +39,7 @@ members = [
     "crates/protos", # Protocol buffer definitions
     "crates/rio", # Rust I/O utilities and abstractions
     "crates/rio-v2", # Next-generation Rust I/O compatibility layer
+    "crates/replication", # Replication contracts and wire formats
     "crates/concurrency", # Concurrency management for RustFS - timeout, locking, backpressure, and I/O scheduling
     "crates/s3-types", # S3 event type definitions
     "crates/s3-ops", # S3 operation definitions and mapping
@@ -106,6 +107,7 @@ rustfs-obs = { path = "crates/obs", version = "1.0.0-beta.8" }
 rustfs-policy = { path = "crates/policy", version = "1.0.0-beta.8" }
 rustfs-protos = { path = "crates/protos", version = "1.0.0-beta.8" }
 rustfs-protocols = { path = "crates/protocols", version = "1.0.0-beta.8" }
+rustfs-replication = { path = "crates/replication", version = "1.0.0-beta.8" }
 rustfs-rio = { path = "crates/rio", version = "1.0.0-beta.8" }
 rustfs-rio-v2 = { path = "crates/rio-v2", version = "1.0.0-beta.8" }
 rustfs-s3-types = { path = "crates/s3-types", version = "1.0.0-beta.8" }
