+++ crates/ecstore/src/bucket/replication/datatypes.rs
@@ -12,36 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use serde::{Deserialize, Serialize};
-use std::fmt;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub enum ResyncStatusType {
-    #[default]
-    NoResync,
-    ResyncPending,
-    ResyncCanceled,
-    ResyncStarted,
-    ResyncCompleted,
-    ResyncFailed,
-}
-
-impl ResyncStatusType {
-    pub fn is_valid(&self) -> bool {
-        *self != ResyncStatusType::NoResync
-    }
-}
-
-impl fmt::Display for ResyncStatusType {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let s = match self {
-            ResyncStatusType::ResyncStarted => "Ongoing",
-            ResyncStatusType::ResyncCompleted => "Completed",
-            ResyncStatusType::ResyncFailed => "Failed",
-            ResyncStatusType::ResyncPending => "Pending",
-            ResyncStatusType::ResyncCanceled => "Canceled",
-            ResyncStatusType::NoResync => "",
-        };
-        write!(f, "{s}")
-    }
-}
+pub use rustfs_replication::ResyncStatusType;
