+++ crates/ecstore/src/bucket/replication/replication_resyncer.rs
@@ -26,6 +26,7 @@ use super::replication_filemeta_boundary::{
 };
 use super::replication_lock_boundary::ReplicationLockTiming;
 use super::replication_metadata_boundary::ReplicationMetadataStore;
+#[cfg(test)]
 use super::replication_msgp_boundary::ReplicationMsgpCodec;
 use super::replication_storage_boundary::{
     AdvancedGetOptions, DeletedObject, EcstoreObjectOperations, HTTPRangeSpec, ObjectInfo, ObjectOptions, ObjectToDelete,
@@ -55,6 +56,7 @@ use http_body::Frame;
 use http_body_util::StreamBody;
 use regex::Regex;
 use rmp_serde;
+use rustfs_replication::{BucketReplicationResyncStatus, ResyncOpts, TargetReplicationResyncStatus};
 use rustfs_s3_types::EventName;
 use rustfs_utils::http::{
     AMZ_BUCKET_REPLICATION_STATUS, AMZ_OBJECT_TAGGING, AMZ_TAGGING_DIRECTIVE, CONTENT_ENCODING, HeaderExt as _,
@@ -73,8 +75,7 @@ use serde::Deserialize;
 use serde::Serialize;
 use std::any::Any;
 use std::collections::HashMap;
-use std::io::{Cursor, Read};
-use std::sync::{Arc, LazyLock};
+use std::sync::Arc;
 use time::OffsetDateTime;
 use time::format_description::well_known::Rfc3339;
 use tokio::io::AsyncRead;
@@ -98,28 +99,13 @@ const EVENT_RESYNC_TASK_FAILED: &str = "replication_resync_task_failed";
 const EVENT_RESYNC_TARGET_OPERATION_FAILED: &str = "replication_resync_target_operation_failed";
 const EVENT_RESYNC_RUNTIME_CHANNEL_FAILED: &str = "replication_resync_runtime_channel_failed";
 
-pub(crate) const RESYNC_META_FORMAT: u16 = 1;
-pub(crate) const RESYNC_META_VERSION: u16 = 1;
+pub(crate) const RESYNC_META_FORMAT: u16 = rustfs_replication::resync::RESYNC_META_FORMAT;
+pub(crate) const RESYNC_META_VERSION: u16 = rustfs_replication::resync::RESYNC_META_VERSION;
 const MRF_META_FORMAT: u16 = 1;
 const MRF_META_VERSION: u16 = 1;
 const RESYNC_TIME_INTERVAL: TokioDuration = TokioDuration::from_secs(60);
-const WIRE_ZERO_TIME_UNIX: i64 = -62_135_596_800;
-
-static WIRE_ZERO_TIME: LazyLock<OffsetDateTime> =
-    LazyLock::new(|| OffsetDateTime::from_unix_timestamp(WIRE_ZERO_TIME_UNIX).unwrap_or(OffsetDateTime::UNIX_EPOCH));
 
 static WARNED_MONITOR_UNINIT: std::sync::Once = std::sync::Once::new();
-
-fn wire_time_or_default(value: Option<OffsetDateTime>) -> OffsetDateTime {
-    value.unwrap_or(*WIRE_ZERO_TIME)
-}
-
-fn normalize_wire_time(value: Option<OffsetDateTime>) -> Option<OffsetDateTime> {
-    match value {
-        Some(v) if v == *WIRE_ZERO_TIME || v == OffsetDateTime::UNIX_EPOCH => None,
-        other => other,
-    }
-}
 
 fn resync_state_accepts_update(state: &TargetReplicationResyncStatus, opts: &ResyncOpts) -> bool {
     state.resync_id.is_empty() || opts.resync_id.is_empty() || state.resync_id == opts.resync_id
@@ -203,152 +189,19 @@ fn content_matches(src: &ObjectInfo, tgt: &HeadObjectOutput) -> bool {
     src_etag.is_some() && src_etag == tgt_etag
 }
 
-#[derive(Debug, Clone, Default)]
-pub struct ResyncOpts {
-    pub bucket: String,
-    pub arn: String,
-    pub resync_id: String,
-    pub resync_before: Option<OffsetDateTime>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct TargetReplicationResyncStatus {
-    pub start_time: Option<OffsetDateTime>,
-    pub last_update: Option<OffsetDateTime>,
-    pub resync_id: String,
-    pub resync_before_date: Option<OffsetDateTime>,
-    pub resync_status: ResyncStatusType,
-    pub failed_size: i64,
-    pub failed_count: i64,
-    pub replicated_size: i64,
-    pub replicated_count: i64,
-    pub bucket: String,
-    pub object: String,
-    pub error: Option<String>,
-}
-
-impl TargetReplicationResyncStatus {
-    pub fn new() -> Self {
-        Self::default()
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct BucketReplicationResyncStatus {
-    pub version: u16,
-    pub targets_map: HashMap<String, TargetReplicationResyncStatus>,
-    pub id: i32,
-    pub last_update: Option<OffsetDateTime>,
-}
-
-impl BucketReplicationResyncStatus {
-    pub fn new() -> Self {
-        Self {
-            version: RESYNC_META_VERSION,
-            ..Default::default()
-        }
-    }
-
-    pub fn clone_tgt_stats(&self) -> HashMap<String, TargetReplicationResyncStatus> {
-        self.targets_map.clone()
-    }
-
-    pub fn marshal_msg(&self) -> Result<Vec<u8>> {
-        let mut wr = Vec::new();
-        rmp::encode::write_map_len(&mut wr, 4)?;
-        rmp::encode::write_str(&mut wr, "v")?;
-        rmp::encode::write_i32(&mut wr, i32::from(self.version))?;
-        rmp::encode::write_str(&mut wr, "brs")?;
-        rmp::encode::write_map_len(&mut wr, self.targets_map.len() as u32)?;
-        for (arn, status) in &self.targets_map {
-            rmp::encode::write_str(&mut wr, arn)?;
-            status.marshal_wire_msg(&mut wr)?;
-        }
-        rmp::encode::write_str(&mut wr, "id")?;
-        rmp::encode::write_i32(&mut wr, self.id)?;
-        rmp::encode::write_str(&mut wr, "lu")?;
-        ReplicationMsgpCodec::write_time(&mut wr, wire_time_or_default(self.last_update))?;
-        Ok(wr)
-    }
-
-    pub fn unmarshal_msg(data: &[u8]) -> Result<Self> {
-        let mut rd = Cursor::new(data);
-        let mut out = Self::new();
-        let mut fields = rmp::decode::read_map_len(&mut rd)?;
-
-        while fields > 0 {
-            fields -= 1;
-            let key = read_msgp_str(&mut rd)?;
-            match key.as_str() {
-                "v" => {
-                    let v: i32 = rmp::decode::read_int(&mut rd)?;
-                    out.version = u16::try_from(v).map_err(|_| Error::other("invalid resync version"))?;
-                }
-                "brs" => {
-                    let map_len = rmp::decode::read_map_len(&mut rd)?;
-                    let mut targets = HashMap::with_capacity(map_len as usize);
-                    for _ in 0..map_len {
-                        let arn = read_msgp_str(&mut rd)?;
-                        let status = TargetReplicationResyncStatus::unmarshal_wire_msg(&mut rd)?;
-                        targets.insert(arn, status);
-                    }
-                    out.targets_map = targets;
-                }
-                "id" => {
-                    out.id = rmp::decode::read_int::<i32, _>(&mut rd)?;
-                }
-                "lu" => {
-                    out.last_update = normalize_wire_time(read_msgp_time_or_nil(&mut rd)?);
-                }
-                _ => ReplicationMsgpCodec::skip_value(&mut rd)?,
-            }
-        }
-        Ok(out)
-    }
-
-    pub fn unmarshal_legacy_msg(data: &[u8]) -> Result<Self> {
-        Ok(rmp_serde::from_slice(data)?)
+fn map_resync_error(err: rustfs_replication::resync::Error) -> Error {
+    match err {
+        rustfs_replication::resync::Error::CorruptedFormat => Error::CorruptedFormat,
+        rustfs_replication::resync::Error::Other(err) => Error::other(err),
     }
 }
 
 pub(crate) fn encode_resync_file(status: &BucketReplicationResyncStatus) -> Result<Vec<u8>> {
-    let payload = status.marshal_msg()?;
-    let mut data = Vec::with_capacity(4 + payload.len());
-    let mut major = [0u8; 2];
-    byteorder::LittleEndian::write_u16(&mut major, RESYNC_META_FORMAT);
-    data.extend_from_slice(&major);
-    let mut minor = [0u8; 2];
-    byteorder::LittleEndian::write_u16(&mut minor, RESYNC_META_VERSION);
-    data.extend_from_slice(&minor);
-    data.extend_from_slice(&payload);
-    Ok(data)
+    rustfs_replication::encode_resync_file(status).map_err(map_resync_error)
 }
 
 pub(crate) fn decode_resync_file(data: &[u8]) -> Result<BucketReplicationResyncStatus> {
-    if data.len() <= 4 {
-        return Err(Error::CorruptedFormat);
-    }
-
-    let mut major = [0u8; 2];
-    major.copy_from_slice(&data[0..2]);
-    if byteorder::LittleEndian::read_u16(&major) != RESYNC_META_FORMAT {
-        return Err(Error::CorruptedFormat);
-    }
-
-    let mut minor = [0u8; 2];
-    minor.copy_from_slice(&data[2..4]);
-    if byteorder::LittleEndian::read_u16(&minor) != RESYNC_META_VERSION {
-        return Err(Error::CorruptedFormat);
-    }
-
-    let status = match BucketReplicationResyncStatus::unmarshal_msg(&data[4..]) {
-        Ok(v) => v,
-        Err(_) => BucketReplicationResyncStatus::unmarshal_legacy_msg(&data[4..])?,
-    };
-    if status.version != RESYNC_META_VERSION {
-        return Err(Error::CorruptedFormat);
-    }
-    Ok(status)
+    rustfs_replication::decode_resync_file(data).map_err(map_resync_error)
 }
 
 pub(crate) fn encode_mrf_file(entries: &[MrfReplicateEntry]) -> Result<Vec<u8>> {
@@ -379,102 +232,6 @@ pub(crate) fn decode_mrf_file(data: &[u8]) -> Result<Vec<MrfReplicateEntry>> {
         return Err(Error::CorruptedFormat);
     }
     rmp_serde::from_slice(&data[4..]).map_err(|e| Error::other(e.to_string()))
-}
-
-impl TargetReplicationResyncStatus {
-    fn marshal_wire_msg(&self, wr: &mut Vec<u8>) -> Result<()> {
-        rmp::encode::write_map_len(wr, 11)?;
-        rmp::encode::write_str(wr, "st")?;
-        ReplicationMsgpCodec::write_time(wr, wire_time_or_default(self.start_time))?;
-        rmp::encode::write_str(wr, "lst")?;
-        ReplicationMsgpCodec::write_time(wr, wire_time_or_default(self.last_update))?;
-        rmp::encode::write_str(wr, "id")?;
-        rmp::encode::write_str(wr, &self.resync_id)?;
-        rmp::encode::write_str(wr, "rdt")?;
-        ReplicationMsgpCodec::write_time(wr, wire_time_or_default(self.resync_before_date))?;
-        rmp::encode::write_str(wr, "rst")?;
-        rmp::encode::write_i32(wr, resync_status_to_i32(self.resync_status))?;
-        rmp::encode::write_str(wr, "fs")?;
-        rmp::encode::write_i64(wr, self.failed_size)?;
-        rmp::encode::write_str(wr, "frc")?;
-        rmp::encode::write_i64(wr, self.failed_count)?;
-        rmp::encode::write_str(wr, "rs")?;
-        rmp::encode::write_i64(wr, self.replicated_size)?;
-        rmp::encode::write_str(wr, "rrc")?;
-        rmp::encode::write_i64(wr, self.replicated_count)?;
-        rmp::encode::write_str(wr, "bkt")?;
-        rmp::encode::write_str(wr, &self.bucket)?;
-        rmp::encode::write_str(wr, "obj")?;
-        rmp::encode::write_str(wr, &self.object)?;
-        Ok(())
-    }
-
-    fn unmarshal_wire_msg<R: Read>(rd: &mut R) -> Result<Self> {
-        let mut out = Self::new();
-        let mut fields = rmp::decode::read_map_len(rd)?;
-
-        while fields > 0 {
-            fields -= 1;
-            let key = read_msgp_str(rd)?;
-            match key.as_str() {
-                "st" => out.start_time = normalize_wire_time(read_msgp_time_or_nil(rd)?),
-                "lst" => out.last_update = normalize_wire_time(read_msgp_time_or_nil(rd)?),
-                "id" => out.resync_id = read_msgp_str(rd)?,
-                "rdt" => out.resync_before_date = normalize_wire_time(read_msgp_time_or_nil(rd)?),
-                "rst" => {
-                    let v: i32 = rmp::decode::read_int(rd)?;
-                    out.resync_status = resync_status_from_i32(v)?;
-                }
-                "fs" => out.failed_size = rmp::decode::read_int(rd)?,
-                "frc" => out.failed_count = rmp::decode::read_int(rd)?,
-                "rs" => out.replicated_size = rmp::decode::read_int(rd)?,
-                "rrc" => out.replicated_count = rmp::decode::read_int(rd)?,
-                "bkt" => out.bucket = read_msgp_str(rd)?,
-                "obj" => out.object = read_msgp_str(rd)?,
-                _ => ReplicationMsgpCodec::skip_value(rd)?,
-            }
-        }
-        Ok(out)
-    }
-}
-
-fn read_msgp_str<R: Read>(rd: &mut R) -> Result<String> {
-    let len = rmp::decode::read_str_len(rd)? as usize;
-    let mut buf = vec![0u8; len];
-    rd.read_exact(&mut buf)?;
-    Ok(String::from_utf8(buf)?)
-}
-
-fn read_msgp_time_or_nil<R: Read>(rd: &mut R) -> Result<Option<OffsetDateTime>> {
-    let marker = rmp::decode::read_marker(rd).map_err(|e| Error::other(format!("{e:?}")))?;
-    match marker {
-        rmp::Marker::Null => Ok(None),
-        rmp::Marker::Ext8 => Ok(Some(ReplicationMsgpCodec::read_ext8_time(rd)?)),
-        other => Err(Error::other(format!("expected time ext or nil, got marker: {other:?}"))),
-    }
-}
-
-fn resync_status_to_i32(status: ResyncStatusType) -> i32 {
-    match status {
-        ResyncStatusType::NoResync => 0,
-        ResyncStatusType::ResyncPending => 1,
-        ResyncStatusType::ResyncCanceled => 2,
-        ResyncStatusType::ResyncStarted => 3,
-        ResyncStatusType::ResyncCompleted => 4,
-        ResyncStatusType::ResyncFailed => 5,
-    }
-}
-
-fn resync_status_from_i32(code: i32) -> Result<ResyncStatusType> {
-    match code {
-        0 => Ok(ResyncStatusType::NoResync),
-        1 => Ok(ResyncStatusType::ResyncPending),
-        2 => Ok(ResyncStatusType::ResyncCanceled),
-        3 => Ok(ResyncStatusType::ResyncStarted),
-        4 => Ok(ResyncStatusType::ResyncCompleted),
-        5 => Ok(ResyncStatusType::ResyncFailed),
-        _ => Err(Error::other(format!("invalid resync status code: {code}"))),
-    }
 }
 
 static RESYNC_WORKER_COUNT: usize = 10;
@@ -4244,7 +4001,8 @@ mod tests {
 
     #[test]
     fn test_resync_none_time_encodes_as_wire_zero_and_decodes_to_none() {
-        let wire_zero = OffsetDateTime::from_unix_timestamp(WIRE_ZERO_TIME_UNIX).expect("valid wire zero timestamp");
+        let wire_zero = OffsetDateTime::from_unix_timestamp(rustfs_replication::resync::WIRE_ZERO_TIME_UNIX)
+            .expect("valid wire zero timestamp");
 
         let mut with_none = BucketReplicationResyncStatus::new();
         with_none.id = 77;
