+++ crates/replication/src/lib.rs
@@ -12,18 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::BucketReplicationResyncStatus;
-use super::replication_error_boundary::Result;
-use super::replication_resyncer::{decode_resync_file, encode_resync_file};
+pub mod resync;
 
-pub(crate) struct ReplicationMigrationBridge;
-
-impl ReplicationMigrationBridge {
-    pub(crate) fn decode_resync_status(data: &[u8]) -> Result<BucketReplicationResyncStatus> {
-        decode_resync_file(data)
-    }
-
-    pub(crate) fn encode_resync_status(status: &BucketReplicationResyncStatus) -> Result<Vec<u8>> {
-        encode_resync_file(status)
-    }
-}
+pub use resync::{
+    BucketReplicationResyncStatus, ResyncOpts, ResyncStatusType, TargetReplicationResyncStatus, decode_resync_file,
+    encode_resync_file,
+};
